requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.12.0
exceptiongroup==1.3.0
fastapi==0.121.2
fastapi-swagger-dark==0.0.8
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyTelegramBotAPI==4.29.1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
ruff==0.14.5
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.44
starlette==0.49.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
win32_setctime==1.2.0
WTForms==3.1.2
yarl==1.22.0
